Added new class for accessing users in the database and changed account sign in to call from this class.
Added AdminController to handle admin management tasks.

Added placeholder view for adding new admins to the database. Not functional. Added method to show this view in ManageController.

Added ability to delete reports.
Added buttons to manage reports in manage index view if user is an admin.
Created reports view to display a grid with article reports. Code for confirmation dialogue box in place but not fumctional.

Updated to latest bootstrap and jquery.